comment route

+++ controllers/post.js
@@ -2,12 +2,13 @@ const DB = require("../models/post");
 const helper = require("../helpers/utils");
 
 const all = async(req, res, next) => {
-    let posts = await DB.find().populate('user', ("-password -__v"));
+    // res.json({msg : "this is all"});
+    let posts = await DB.find().populate("user cat", "-__v -password");
     helper.fMmsg(res, "All posts", posts);
 }
 
 const get = async(req, res, next) => {
-     let post = await DB.findById(req.params.id).populate("user", "-password -__v");
+     let post = await DB.findById(req.params.id).populate("user cat", "-password -__v");
      if(post){
         helper.fMmsg(res, "Single Post", post);
      }else{
@@ -16,9 +17,12 @@ const get = async(req, res, next) => {
 }
 
 const post = async(req, res, next) => {
-    // console.log(req.body.user);
-    // let savePost = await new DB(req.body).save();
-    helper.fMmsg(res, "post Added", req.body.user);
+    // console.log(req.body);
+    let userId  = req.body.user._id;
+    delete req.body.user;
+    req.body.user = userId;
+    let savePost = await new DB(req.body).save();
+    helper.fMmsg(res, "post Added", savePost);
 }
 
 const patch = async(req, res, next) => {
@@ -32,9 +36,25 @@ const patch = async(req, res, next) => {
     }
 }
 
+const bycat = async(req, res, next) => {
+    let posts = await DB.find({cat: req.params.id});
+    // console.log(posts);
+    helper.fMmsg(res, "all post from cat", posts);
+}
+
+const byuserId = async(req, res, next) => {
+    let post = await DB.find({user: req.params.id}).populate("user", "-password -__v");
+    helper.fMmsg(res, "all post from user", post);
+}
+
 const drop = async(req, res, next) => {
-    await DB.findByIdAndDelete(req.params.id);
-    helper.fMmsg(res, "Post Deleted");
+    let post = await DB.findById(req.params.id);
+    if(post){
+        await DB.findByIdAndDelete(req.params.id);
+        helper.fMmsg(res, "Post Deleted");
+    }else{
+        next(new Error("No post is with that id"));
+    }
 }
 
 module.exports = {
@@ -42,5 +62,7 @@ module.exports = {
     get,
     post,
     patch,
-    drop
+    drop,
+    bycat,
+    byuserId
 }
